Update branding of WebAuthn RP name
`rp.name` is a human-friendly label for the web service requesting
WebAuthn credential creation. The browser may show this label to the
user during credential creation and/or in the authenticator's
credential management interface (few or no browsers currently do,
though - most just show `rp.id` which is the service's domain name).