Move additional base64 encoding of WebAuthn responses to backend
We already have JSON body parsers configured to automatically wrap and unwrap
binary data, so we don't need this encoding for the frontend's sake. That the
backend uses SimpleWebauthn which expects inputs base64url encoded is an
implementation detail that should not leak into the API exposed to the frontend.