Ignore attestation verification errors during signup
This removes the attestation statement from the
`SimpleWebauthn.verifyRegistrationResponse` call, so that for example expired
attestation certs don't cause the registration to fail. We only want the
attestation for informational purposes, such as being able to monitor
vulnerability reports and warn affected users; we don't actually care whether
the attestation is valid.