Recognize If-None-Match header in GET /session/private-data
If `GET /session/private-data` is called with an [`If-None-Match`][1] or
`X-Private-Data-If-None-Match` header, then the request is a [conditional
request][2] and will return [304 Not Modified][3] if the header value matches
the ETag value that would be returned in a 200 OK response. This enables clients
to atomically update their local copy of the private data only if needed, while
not re-sending another copy of the same private data when not needed.

[1]: https://www.rfc-editor.org/rfc/rfc9110.html#name-if-none-match
[2]: https://www.rfc-editor.org/rfc/rfc9110.html#name-conditional-requests
[3]: https://www.rfc-editor.org/rfc/rfc9110.html#name-304-not-modified